app/backend/pyproject.toml
[project]
name = "met-art-search-backend"
version = "0.1.0"
description = "MET Art Search FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "sentence-transformers>=3.0.0",
    "Pillow>=10.3.0",
    "pydantic-settings>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
